requirements.txt
pandas==2.0.3
numpy==1.24.3
cassandra-driver==3.28.0
requests==2.31.0
streamlit==1.25.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
scikit-learn==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
schedule==1.2.0
flask==2.3.2
gunicorn==21.2.0